% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeostTextureLibrarySharedRoxy.R
\name{anisoDirL}
\alias{anisoDirL}
\title{Calculate the direction of maximum continuity considering 4 directions}
\usage{
anisoDirL(x)
}
\arguments{
\item{x}{A list of rasters with the spatial variability along 4 directions (see function anisoDir())}
}
\value{
A raster with the direction (in degrees, geographical) of maximum continuity
}
\description{
The input is represented by a list of rasters with the spatial variability index (e.g., MAD, variogram, etc.) computed
in four directions (N-S, NE-SW, E-W, SE-NW)
}
