% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeostTextureLibrarySharedRoxy.R
\name{anisoR}
\alias{anisoR}
\title{Calculate the index of anisotropy considering the spatial variability along 4 directions}
\usage{
anisoR(N, NE, E, SE)
}
\arguments{
\item{N}{Spatial vairability along N-S direction}

\item{NE}{Spatial vairability along NE-SW direction}

\item{E}{Spatial vairability along E-W direction}

\item{SE}{Spatial vairability along SE-NW direction}
}
\value{
A raster with the index of anisotropy (min=0 max=1)
}
\description{
The input is represented by four rasters with the spatial variability index (e.g., MAD, variogram, etc.) computed
in four directions (N-S, NE-SW, E-W, SE-NW)
}
