% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeostTextureLibrarySharedRoxy.R
\name{anisoRL}
\alias{anisoRL}
\title{Calculate the index of anisotropy considering the spatial variability along 4 directions}
\usage{
anisoRL(x)
}
\arguments{
\item{x}{A list of rasters with the spatial variability along 4 directions (see function anisoR())}
}
\value{
A raster with the index of anisotropy (min=0 max=1)
}
\description{
The input is represented by a list of rasters with the spatial variability index (e.g., MAD, variogram, etc.) computed
in four directions (N-S, NE-SW, E-W, SE-NW)
}
