% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeostTextureLibrarySharedRoxy.R
\name{CalcMeans}
\alias{CalcMeans}
\title{Calculate the mean of absolute values raised to an exponent found in a search window}
\usage{
CalcMeans(deltas, w, exponent)
}
\arguments{
\item{deltas}{The values from which calculate the median of absolute values (i.e., directional differences of order K)}

\item{w}{The moving window used (e.g. w=KernelCircular(3))}

\item{exponent}{The exponent: increasing the exponent increase the sensitivity to outliers. Set 2 for Variogram and 1 for Madogram.}
}
\value{
A raster with the mean of absolute values in the search window
}
\description{
With this you can compute variogram and madogram (but remember that for
classical geostatistical indexes you need to divide the derived isotropic index by 2!)
}
