% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeostTextureLibrarySharedRoxy.R
\name{CalcMedians}
\alias{CalcMedians}
\title{Calculate the median of absolute values found in a search window for each raster in a list}
\usage{
CalcMedians(deltas, w)
}
\arguments{
\item{deltas}{A list of rasters with the values from which calculate the median of absolute values (e.g., directional differences of order K)}

\item{w}{The moving window used (e.g. w=KernelCircular(3))}
}
\value{
A list of rasters with the median of absolute values in the search window
}
\description{
Calculate the median of absolute values found in a search window for each raster in a list
}
