% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeostTextureLibrarySharedRoxy.R
\name{circularEigenNV}
\alias{circularEigenNV}
\title{Compute circular variance of normal vectors to surface}
\usage{
circularEigenNV(inraster, window)
}
\arguments{
\item{inraster}{The DEM from which compute the index}

\item{window}{The moving window adopted for computing the index}
}
\value{
The raster with the computed index
}
\description{
Compute circular variance of normal vectors to surface, using the eigen values (only for testing, very slow)
}
