% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{k05ck2}
\alias{k05ck2}
\title{basic kernels}
\format{
just matrices.
}
\source{
Sebastiano Trevisani
}
\usage{
k05ck2
}
\description{
Kernels for computing directional differences for specific directions and lag distances.
These have been constructed using bilinear interpolation for directions out of main axes.
The kernels are intended to be used with "Terra" focal functions (i.e., convolution).
}
\examples{
#to see kernels (each one is a list with 4 kernels) of order 1
#These should be used with a detrended "surface"
#lag 1 pixel
k1c
#lag 2 pixels
k2c
#lag 4 pixels
k4c
#lag 6 pixels
k6c
#lag 8 pixels
k8c
#kernels of order 2 (differences of differences)
#these can be applied directly without detrending
#lag 05 pixel
k05ck2
#lag 1 pixel
k1ck2
#lag 2 pixels
k2ck2
}
\references{
xxx
}
\keyword{datasets}
