% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeostTextureLibrarySharedRoxy.R
\name{KernelCircular}
\alias{KernelCircular}
\title{Build a circular moving window}
\usage{
KernelCircular(radius)
}
\arguments{
\item{radius}{The radius of the moving window}
}
\value{
A matrix with selected pixels
}
\description{
Build a circular moving window
}
\examples{
#A circular moving window with a radius of 3 pixels
w=KernelCircular(3)
w
}
