% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeostTextureLibrarySharedRoxy.R
\name{matDeg}
\alias{matDeg}
\title{Conversion from geographical directions to mathematical ones}
\usage{
matDeg(alpha)
}
\arguments{
\item{alpha}{The angle in degrees from one system of representation (e.g. geographical)}
}
\value{
The angle in degrees in the other system
}
\description{
Conversion from geographical directions to mathematical ones
}
\examples{
#a=matDeg(45)
#a
}
