% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeostTextureLibrarySharedRoxy.R
\name{rad}
\alias{rad}
\title{Conversion to radians}
\usage{
rad(degree)
}
\arguments{
\item{degree}{Angle in degrees}
}
\value{
The angle in radians
}
\description{
Conversion to radians
}
\examples{
#rad(45)
}
