% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeostTextureLibrarySharedRoxy.R
\name{roory}
\alias{roory}
\title{For computing vector dispersion using eigen values ratios}
\usage{
roory(x)
}
\arguments{
\item{x}{Matrix cross products}
}
\value{
The dispersion/smoothness
}
\description{
For computing vector dispersion using eigen values ratios
}
\examples{
#
}
